requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-socketio[asgi]==5.9.0